src/app/modules/valorisation-organique-economie-circulaire/ai-model/notebooks/requirements.txt
# Python >= 3.10
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
PyYAML>=6.0.0
onnx>=1.15.0
numpy>=1.24.0
tqdm>=4.66.0

onnxruntime>=1.16.0
